seperate pdf export util & added export count to status
