+++ app.py
@@ -7,25 +7,17 @@ import time
 import re
 import sys
 import logging
-import io
-from reportlab.lib.pagesizes import letter
-from reportlab.pdfgen import canvas
 from slack_bolt import App
 from slack_bolt.adapter.socket_mode import SocketModeHandler
 from slack_bolt.authorization import AuthorizeResult
 from slack_sdk import WebClient
 from slack_sdk.errors import SlackApiError
-from reportlab.lib.pagesizes import letter
-from reportlab.lib.styles import getSampleStyleSheet, ParagraphStyle
-from reportlab.platypus import SimpleDocTemplate, Paragraph, Spacer, ListFlowable, ListItem
 import io
 from utils.slack_api import send_message
 from chains.chat_chain_mcp import process_message_mcp, _get_memory, _memories
 from chains.analyze_thread import analyze_slack_thread
 from utils.slack_tools import get_user_name
-
-logging.basicConfig(level=logging.DEBUG)
-
+from utils.export_pdf import render_summary_to_pdf
 SLACK_APP_TOKEN      = os.getenv("SLACK_APP_TOKEN")
 SLACK_SIGNING_SECRET = os.getenv("SLACK_SIGNING_SECRET")
 BOT_USER_ID          = os.getenv("BOT_USER_ID")
@@ -94,6 +86,7 @@ def load_stats():
             "analyze_followups": d.get("analyze_followups", 0),
             "general_calls": d.get("general_calls", 0),
             "general_followups": d.get("general_followups", 0),
+            "pdf_exports": d.get("pdf_exports", 0),
         }
     except:
         return {
@@ -105,6 +98,7 @@ def load_stats():
             "analyze_followups": 0,
             "general_calls": 0,
             "general_followups": 0,
+            "pdf_exports": 0,
         }
 
 def save_stats():
@@ -144,6 +138,7 @@ USAGE_STATS = {
     "analyze_followups": _stats["analyze_followups"],
     "general_calls": _stats["general_calls"],
     "general_followups": _stats["general_followups"],
+    "pdf_exports": _stats["pdf_exports"],  # NEW: track PDF exports
 }
 
 # NEW: track which threads began as an analysis
@@ -177,67 +172,6 @@ def resolve_user_mentions(client: WebClient, text: str) -> str:
         text,
     )
     return text
-def render_summary_to_pdf(text: str) -> io.BytesIO:
-    buffer = io.BytesIO()
-    doc = SimpleDocTemplate(
-        buffer,
-        pagesize=letter,
-        rightMargin=40, leftMargin=40,
-        topMargin=40, bottomMargin=40,
-    )
-
-    styles = getSampleStyleSheet()
-    styles.add(ParagraphStyle(
-        name='Heading',
-        parent=styles['Heading2'],
-        fontSize=14,
-        leading=18,
-        fontName='Helvetica-Bold'
-    ))
-    styles.add(ParagraphStyle(
-        name='Body',
-        parent=styles['BodyText'],
-        fontSize=10,
-        leading=12
-    ))
-
-    flowables = []
-    for line in text.splitlines():
-        line = line.strip()
-        if not line:
-            continue
-
-        # 1) Detect "*Heading*" at start, possibly with trailing text
-        m = re.match(r'^\*(.+?)\*\s*(.*)$', line)
-        if m:
-            heading = m.group(1).strip()
-            rest    = m.group(2).strip()
-            flowables.append(Paragraph(heading, styles['Heading']))
-            flowables.append(Spacer(1, 6))
-            if rest:
-                flowables.append(Paragraph(rest, styles['Body']))
-                flowables.append(Spacer(1, 6))
-            continue
-
-        # 2) Bullet list item
-        if line.startswith('- '):
-            # collect consecutive bullets into one ListFlowable
-            items = []
-            # note: we only have one line at a time here,
-            # so for simplicity treat each "- " line as its own list
-            item = line[2:].strip()
-            items.append(ListItem(Paragraph(item, styles['Body']), leftIndent=12))
-            flowables.append(ListFlowable(items, bulletType='bullet'))
-            flowables.append(Spacer(1, 6))
-            continue
-
-        # 3) Regular paragraph
-        flowables.append(Paragraph(line, styles['Body']))
-        flowables.append(Spacer(1, 6))
-
-    doc.build(flowables)
-    buffer.seek(0)
-    return buffer
 @app.action("export_pdf")
 def handle_export_pdf(ack, body, client, logger):
     ack()
@@ -251,7 +185,7 @@ def handle_export_pdf(ack, body, client, logger):
     plain = re.sub(r'\r\n?', '\n', summary_md)
 
     pdf_buffer = render_summary_to_pdf(plain)
-
+    USAGE_STATS["pdf_exports"] += 1
     client.files_upload_v2(
         channels=[channel_id],
         file=pdf_buffer,
@@ -301,7 +235,8 @@ def get_bot_stats():
         "📊 *Bot Usage Stats*\n"
         f"• *Total calls:* {USAGE_STATS['total_calls']}\n"
         f"• *Analyze calls:* {USAGE_STATS['analyze_calls']} (follow-ups: {USAGE_STATS['analyze_followups']})\n"
-        f"• *General calls:* {USAGE_STATS['general_calls']} (follow-ups: {USAGE_STATS['general_followups']})\n\n"
+        f"• *General calls:* {USAGE_STATS['general_calls']} (follow-ups: {USAGE_STATS['general_followups']})\n"
+        f"• *PDF exports:* {USAGE_STATS['pdf_exports']}\n\n"
         f"👍 *{_vote_up_count}*   👎 *{_vote_down_count}*"
     )
