+++ requirements.txt
@@ -3,3 +3,4 @@ langchain-community
 slack-bolt
 python-dotenv
 requests
+reportlab
